generalize pkg/network/stats Reporter for broader usage
