+++ pkg/network/http/telemetry.go
@@ -9,10 +9,9 @@
 package http
 
 import (
-	"fmt"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/network/stats"
+	"github.com/DataDog/datadog-agent/pkg/util/atomicstats"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"go.uber.org/atomic"
 )
@@ -27,9 +26,9 @@ type telemetry struct {
 	rejected                                    *atomic.Int64 `stats:""` // this happens when an user-defined reject-filter matches a request
 	malformed                                   *atomic.Int64 `stats:""` // this happens when the request doesn't have the expected format
 	aggregations                                *atomic.Int64 `stats:""`
-
-	reporter stats.Reporter
 }
+
+var telemetryReporter = atomicstats.NewReporter((*telemetry)(nil))
 
 func newTelemetry() (*telemetry, error) {
 	t := &telemetry{
@@ -45,12 +44,6 @@ func newTelemetry() (*telemetry, error) {
 		rejected:     atomic.NewInt64(0),
 		malformed:    atomic.NewInt64(0),
 		aggregations: atomic.NewInt64(0),
-	}
-
-	var err error
-	t.reporter, err = stats.NewReporter(t)
-	if err != nil {
-		return nil, fmt.Errorf("error creating stats reporter: %w", err)
 	}
 
 	return t, nil
@@ -114,5 +107,5 @@ func (t *telemetry) reset() telemetry {
 }
 
 func (t *telemetry) report() map[string]interface{} {
-	return t.reporter.Report()
+	return telemetryReporter.Report(t)
 }
