+++ pkg/network/tracer/connection/kprobe/tracer.go
@@ -20,9 +20,9 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	netebpf "github.com/DataDog/datadog-agent/pkg/network/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/network/ebpf/probes"
-	"github.com/DataDog/datadog-agent/pkg/network/stats"
 	"github.com/DataDog/datadog-agent/pkg/network/tracer/connection"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
+	"github.com/DataDog/datadog-agent/pkg/util/atomicstats"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	manager "github.com/DataDog/ebpf-manager"
 	"github.com/cilium/ebpf"
@@ -55,9 +55,9 @@ type kprobeTracer struct {
 type telemetry struct {
 	tcpConns4, tcpConns6 *atomic.Int64 `stats:""`
 	udpConns4, udpConns6 *atomic.Int64 `stats:""`
-
-	reporter stats.Reporter
 }
+
+var telemetryReporter = atomicstats.NewReporter((*telemetry)(nil))
 
 func newTelemetry() telemetry {
 	return telemetry{
@@ -175,11 +175,6 @@ func New(config *config.Config, constants []manager.ConstantEditor) (connection.
 	if err != nil {
 		tr.Stop()
 		return nil, fmt.Errorf("error retrieving the bpf %s map: %s", probes.TcpStatsMap, err)
-	}
-
-	tr.telemetry.reporter, err = stats.NewReporter(&tr.telemetry)
-	if err != nil {
-		return nil, fmt.Errorf("error creating stats reporter: %w", err)
 	}
 
 	return tr, nil
@@ -302,7 +297,7 @@ func (t *telemetry) removeConnection(conn *network.ConnectionStats) {
 }
 
 func (t *telemetry) get() map[string]interface{} {
-	return t.reporter.Report()
+	return telemetryReporter.Report(t)
 }
 
 func (t *kprobeTracer) Remove(conn *network.ConnectionStats) error {
