+++ pkg/network/tracer/gateway_lookup.go
@@ -16,8 +16,8 @@ import (
 	ddconfig "github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/network"
 	"github.com/DataDog/datadog-agent/pkg/network/config"
-	"github.com/DataDog/datadog-agent/pkg/network/stats"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
+	"github.com/DataDog/datadog-agent/pkg/util/atomicstats"
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/hashicorp/golang-lru/simplelru"
@@ -39,9 +39,9 @@ type gatewayLookup struct {
 	subnetCacheLookups *atomic.Uint64 `stats:""`
 	subnetLookups      *atomic.Uint64 `stats:""`
 	subnetLookupErrors *atomic.Uint64 `stats:""`
-
-	statsReporter stats.Reporter
 }
+
+var gatewayLookupReporter = atomicstats.NewReporter((*gatewayLookup)(nil))
 
 type cloudProvider interface {
 	IsAWS() bool
@@ -71,13 +71,6 @@ func newGatewayLookup(config *config.Config) *gatewayLookup {
 		subnetCacheLookups:  atomic.NewUint64(0),
 		subnetLookups:       atomic.NewUint64(0),
 		subnetLookupErrors:  atomic.NewUint64(0),
-	}
-
-	var err error
-	gl.statsReporter, err = stats.NewReporter(gl)
-	if err != nil {
-		log.Errorf("could not create stats reporter for gateway lookup: %s", err)
-		return nil
 	}
 
 	router, err := network.NewNetlinkRouter(config.ProcRoot)
@@ -186,7 +179,7 @@ func (g *gatewayLookup) GetStats() map[string]interface{} {
 		return make(map[string]interface{})
 	}
 
-	report := g.statsReporter.Report()
+	report := gatewayLookupReporter.Report(g)
 	report["route_cache"] = g.routeCache.GetStats()
 	return report
 }
