+++ pkg/util/atomicstats/reporter_test.go
@@ -1,18 +1,61 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2022-present Datadog, Inc.
 
-package stats
+package atomicstats
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"go.uber.org/atomic"
 )
 
-func TestStatsAllowedTypes(t *testing.T) {
+func Example() {
+	// define a struct with some `stats` tags
+	type myStats struct {
+		integer       int64         `stats:""`
+		atomicInteger *atomic.Int64 `stats:""`
+		notStats      int64
+	}
+
+	// create a singleton reporter for this type
+	myStatsReporter := NewReporter((*myStats)(nil))
+
+	// create a myStats value
+	stats := myStats{
+		integer:       10,
+		atomicInteger: atomic.NewInt64(20),
+		notStats:      30,
+	}
+	statsMap := myStatsReporter.Report(&stats)
+
+	fmt.Printf("%#v\n", statsMap)
+	// Output:
+	// map[string]interface {}{"atomic_integer":20, "integer":10}
+}
+
+func TestNewReporter_NotPtr(t *testing.T) {
+	type myStats struct{}
+	require.Panics(t, func() { NewReporter(myStats{}) })
+}
+
+func TestNewReporter_NotStructPtr(t *testing.T) {
+	someNumber := 13
+	require.Panics(t, func() { NewReporter(&someNumber) })
+}
+
+func TestNewReporter_BadType(t *testing.T) {
+	type myStats struct {
+		// (if and when we support strings, think of something more interesting)
+		stringStat string `stats:""`
+	}
+	require.Panics(t, func() { NewReporter((*myStats)(nil)) })
+}
+
+func TestReporterAllowedTypes(t *testing.T) {
 	//nolint:structcheck,unused
 	type test struct {
 		i64 int64 `stats:""`
@@ -30,10 +73,8 @@ func TestStatsAllowedTypes(t *testing.T) {
 		uptr uintptr `stats:""`
 	}
 
-	v := &test{}
-	s, err := NewReporter(v)
-	require.NoError(t, err)
-	stats := s.Report()
+	s := NewReporter((*test)(nil))
+	stats := s.Report(&test{})
 	require.Len(t, stats, 11)
 	require.Equal(t, int64(0), stats["i64"])
 	require.Equal(t, int32(0), stats["i32"])
@@ -50,7 +91,7 @@ func TestStatsAllowedTypes(t *testing.T) {
 	require.Equal(t, uintptr(0), stats["uptr"])
 }
 
-func TestStatsSnakeCase(t *testing.T) {
+func TestReporterSnakeCase(t *testing.T) {
 	//nolint:structcheck,unused
 	type test struct {
 		foo       int `stats:""`
@@ -58,10 +99,8 @@ func TestStatsSnakeCase(t *testing.T) {
 		barbaz    int `stats:""`
 		fooBarBaz int `stats:""`
 	}
-
-	s, err := NewReporter(&test{})
-	require.NoError(t, err)
-	stats := s.Report()
+	s := NewReporter((*test)(nil))
+	stats := s.Report(&test{})
 	require.Len(t, stats, 4)
 	require.Contains(t, stats, "foo")
 	require.Contains(t, stats, "bar_baz")
@@ -69,7 +108,7 @@ func TestStatsSnakeCase(t *testing.T) {
 	require.Contains(t, stats, "foo_bar_baz")
 }
 
-func TestStatsSkipNoTag(t *testing.T) {
+func TestReporterSkipNoTag(t *testing.T) {
 	//nolint:structcheck,unused
 	type test struct {
 		foo int `stats:""`
@@ -77,17 +116,15 @@ func TestStatsSkipNoTag(t *testing.T) {
 		baz int `stats:""`
 	}
 
-	v := &test{}
-	s, err := NewReporter(v)
-	require.NoError(t, err)
-	stats := s.Report()
+	s := NewReporter((*test)(nil))
+	stats := s.Report(&test{})
 	require.Len(t, stats, 2)
 	require.Contains(t, stats, "foo")
 	require.Contains(t, stats, "baz")
 	require.NotContains(t, stats, "bar")
 }
 
-func TestStatsAllowedTypesAtomic(t *testing.T) {
+func TestReporterAllowedTypesAtomic(t *testing.T) {
 	//nolint:structcheck,unused
 	type test struct {
 		boolp *atomic.Bool   `stats:""`
@@ -95,14 +132,13 @@ func TestStatsAllowedTypesAtomic(t *testing.T) {
 		u64p  *atomic.Uint64 `stats:""`
 	}
 
+	s := NewReporter((*test)(nil))
 	v := &test{
 		boolp: atomic.NewBool(true),
 		i64p:  atomic.NewInt64(6),
 		u64p:  atomic.NewUint64(7),
 	}
-	s, err := NewReporter(v)
-	require.NoError(t, err)
-	stats := s.Report()
+	stats := s.Report(v)
 	require.Len(t, stats, 3)
 	require.Equal(t, true, stats["boolp"])
 	require.Equal(t, int64(6), stats["i64p"])
